+++ parse/map_string_string_slice_test.go
@@ -1,4 +1,4 @@
-package parsestring
+package parse
 
 import (
 	"fmt"
