+++ parse/split_string_slice.go
@@ -1,4 +1,4 @@
-package parsestring
+package parse
 
 import (
 	"fmt"
@@ -7,27 +7,27 @@ import (
 	"text/scanner"
 )
 
-// splitMap splits the values that are csv key:value-pairs of strings in a
-// string into a map[string][]string.
-func splitMap(s string, addKV func(k, v string) error) error {
+// splitStringsSlice splits up a string composed of comma-separated values.
+// Destination type determined by func passed in.
+func splitStringsSlice(s string, addVal func(val string) error) error {
+	if len(s) == 0 {
+		return nil
+	}
 	errs := map[scanner.Position]string{}
+
+	inValue := true
 
 	// initialize the scanner (note: sc.Init() blindly overwrites fields with sane-defaults)
 	sc := scanner.Scanner{}
 	sc.Init(strings.NewReader(s))
-	// We need different defaults for Mode and Error, though
+	// We need a different default for Mode and Error, though
 	sc.Mode = scanner.ScanStrings | scanner.ScanRawStrings |
 		scanner.ScanIdents | scanner.ScanInts |
 		scanner.ScanFloats | scanner.ScanChars
 	sc.Error = func(s *scanner.Scanner, msg string) {
 		errs[s.Pos()] = msg
 	}
-
-	inKey := true
-	inValue := false
-	curKey := ""
-	curVal := ""
-	for tok := sc.Scan(); sc.ErrorCount == 0; tok = sc.Scan() {
+	for tok := sc.Scan(); tok != scanner.EOF && sc.ErrorCount == 0; tok = sc.Scan() {
 		switch tok {
 		case scanner.String, scanner.RawString, scanner.Ident, scanner.Float, scanner.Int:
 			txt := sc.TokenText()
@@ -38,41 +38,21 @@ func splitMap(s string, addKV func(k, v string) error) error {
 				}
 				txt = parsedtxt
 			}
-
-			if inKey {
-				curKey = txt
-			} else if inValue && curKey != "" {
-				curVal = txt
+			if inValue {
+				if addErr := addVal(txt); addErr != nil {
+					return fmt.Errorf("failed to add val %q: %s",
+						txt, addErr)
+				}
 			} else {
 				return fmt.Errorf("unexpected string literal: %s",
 					sc.TokenText())
 			}
-		case ',':
-			if curKey != "" {
-				if addErr := addKV(curKey, curVal); addErr != nil {
-					return fmt.Errorf("map parsing failed on key %q: %s",
-						curKey, addErr)
-				}
-			}
-
-			curKey = ""
-			curVal = ""
-			inKey = true
 			inValue = false
-		case ':':
-			if inValue || curKey == "" {
-				return fmt.Errorf("unexpected colon")
-			}
-			inKey = false
+		case ',':
 			inValue = true
-		case scanner.EOF:
-			if curKey != "" {
-				if addErr := addKV(curKey, curVal); addErr != nil {
-					return fmt.Errorf("map parsing failed on key %q: %s",
-						curKey, addErr)
-				}
-			}
-			return nil
+		default:
+			return fmt.Errorf("unexpected token %s with value: %q",
+				scanner.TokenString(tok), sc.TokenText())
 		}
 	}
